Add the operation forms (update category on select change, create and delete)

+++ src/main/java/com/perso/bank/domain/dto/CategoryDTO.java
@@ -19,6 +19,6 @@ public class CategoryDTO {
 	
 	private String name;
 	
-	private List<Integer> operations;
+	private List<Long> operations;
 	
 }
